Adding PlayerPrefs to save data
Created new class to handle user input data, created instance for that class to be able to access that information.
Created PlayerPrefs to save data to file and be able to use it between scenes/sessions. Added functionality that saves data to file only if it is a new high score.